+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@ This version has some breaking changes:
 
 * The `DeferredHint` protocol has been removed
 * The `DeferredTutorial` protocol has been removed
+* The `LocalizedTutorial` class has been removed
 * `StandardDeferredHint` has been renamed to `DeferredHint`
 * `StandardDeferredTutorial` has been renamed to `DeferredTutorial`